#4 minter refactoring
