+++ test/DividendMinter.js
@@ -1,11 +1,11 @@
 const { ethers, upgrades } = require('hardhat')
 const { expect } = require('chai')
 
-describe('DividendTokenMinter', function () {
+describe('DividendMinter', function () {
   before(async function () {
     this.ProfitToken = await ethers.getContractFactory('ProfitToken')
     this.DividendToken = await ethers.getContractFactory('DividendToken')
-    this.DividendMinter = await ethers.getContractFactory('DividendTokenMinter')
+    this.DividendMinter = await ethers.getContractFactory('DividendMinter')
     const [devFund, tester] = await ethers.getSigners()
     this.devFund = devFund
     this.tester = tester
@@ -25,15 +25,9 @@ describe('DividendTokenMinter', function () {
 
     await this.dividendToken.deployed()
 
-    this.dividendTokenMinter = await upgrades.deployProxy(
+    this.dividendMinter = await upgrades.deployProxy(
       this.DividendMinter,
-      [
-        this.token.address,
-        this.dividendToken.address,
-        1,
-        100000,
-        this.devFund.address,
-      ],
+      [this.token.address, this.dividendToken.address, 1, 100000],
       {
         kind: 'uups',
       }
@@ -41,10 +35,8 @@ describe('DividendTokenMinter', function () {
   })
 
   it('deployed', async function () {
-    expect(await this.dividendTokenMinter.stakeToken()).to.equal(
-      this.token.address
-    )
-    expect(await this.dividendTokenMinter.rewardToken()).to.equal(
+    expect(await this.dividendMinter.stakeToken()).to.equal(this.token.address)
+    expect(await this.dividendMinter.rewardToken()).to.equal(
       this.dividendToken.address
     )
   })
@@ -52,11 +44,11 @@ describe('DividendTokenMinter', function () {
   it('stake, update', async function () {
     await this.token
       .connect(this.devFund)
-      .approve(this.dividendTokenMinter.address, 10)
+      .approve(this.dividendMinter.address, 10)
 
-    await expect(this.dividendTokenMinter.connect(this.devFund).stake(9)).to.not
-      .be.reverted
+    await expect(this.dividendMinter.connect(this.devFund).stake(9)).to.not.be
+      .reverted
 
-    await expect(this.dividendTokenMinter.update()).to.not.be.reverted
+    await expect(this.dividendMinter.update()).to.not.be.reverted
   })
 })
